requirements.txt
boto3
line-bot-sdk
openai
